pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "setuptools_scm[toml]>=6.2", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aigor"
dynamic = ["version"]
description = "AIgor is an assistant for oldshol unix users."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  'typer',
  'rich',
  'python-dotenv',
  'anthropic',
]
authors = [
  {'name' = 'Igor Morgado', 'email' = "morgado.igor@gmail.com"},
]
maintainers = [
  {'name' = 'Igor Morgado', 'email' = "morgado.igor@gmail.com"},
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pylint",
  "mypy",
  "ipython",
  "ipdb",
  "pytest-mypy", 
  "setuptools-scm",
  "types-PyYAML",
]

[tool.setuptools.dynamic]
version = {attr = "aigor.__version__"}

[project.scripts]
aigor = "aigor.main:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
aigor = ["py.typed"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --mypy"
testpaths = [
    "tests",
    # "integration",
]

[tool.pytest]
pythonpath = "."

[tool.mypy]
ignore_missing_imports = true
strict = true

[tool.pylint.format]
max-line-length = "88"

[tool.pylint.design]
max-args = "5"
max-attributes = "7"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.py$'
